database: update for new schema changes
The algorithms now have parameters, that get statically checked before the insert by the database. Also some other small tweaks necessary for the AntiSpoof Algorithm in the Worker-Service. the Table Answer now has a qualityAssured bit, that tells the WS to trust the quality and prohibit answers that are spoof/duplicates from being rated. Also added the stopgap functionality