update to the comments
The longs story:

for the future the database want to give a platform the possibility to
add json as identification field. The platform code can then perform
faster queriers on the dataset than the normal searching. This is why we
have a workerIdentification defualt implementation which does only
search the worker with the correct uid.